front: project trains only in debug mode in stdcm
